requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2024.12.14
decorator==5.1.1
distro==1.9.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
jiter==0.8.2
moviepy==2.1.2
numpy==2.2.1
openai==1.59.7
pillow==10.4.0
proglog==0.1.10
pydantic==2.10.5
pydantic_core==2.27.2
pydub==0.25.1
python-dotenv==1.0.1
pytube==15.0.0
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
yt-dlp==2025.1.12
